services/pdf-processor/requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
requests-mock>=1.11.0

# Testing utilities
httpx>=0.24.0
pytest-asyncio>=0.21.0
